+++ lib/src/analyzers/lint_analyzer/rules/rules_list/check_for_equals_in_render_object_setters/check_for_equals_in_render_object_setters_rule.dart
@@ -1,0 +1,43 @@
+// ignore_for_file: public_member_api_docs
+
+import 'package:analyzer/dart/ast/ast.dart';
+import 'package:analyzer/dart/ast/visitor.dart';
+import 'package:analyzer/dart/element/element.dart';
+
+import '../../../../../utils/flutter_types_utils.dart';
+import '../../../../../utils/node_utils.dart';
+import '../../../lint_utils.dart';
+import '../../../models/internal_resolved_unit_result.dart';
+import '../../../models/issue.dart';
+import '../../../models/severity.dart';
+import '../../models/flutter_rule.dart';
+import '../../rule_utils.dart';
+
+part 'visitor.dart';
+
+class CheckForEqualsInRenderObjectSettersRule extends FlutterRule {
+  static const ruleId = 'check-for-equals-in-render-object-setters';
+
+  CheckForEqualsInRenderObjectSettersRule([
+    Map<String, Object> config = const {},
+  ]) : super(
+          id: ruleId,
+          severity: readSeverity(config, Severity.warning),
+          excludes: readExcludes(config),
+        );
+
+  @override
+  Iterable<Issue> check(InternalResolvedUnitResult source) {
+    final visitor = _Visitor();
+
+    source.unit.visitChildren(visitor);
+
+    return visitor.declarations
+        .map((declaration) => createIssue(
+              rule: this,
+              location: nodeLocation(node: declaration.node, source: source),
+              message: declaration.errorMessage,
+            ))
+        .toList(growable: false);
+  }
+}